refactor error msg
